requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cfgv==3.3.1
distlib==0.3.5
Django==4.0.6
django-ckeditor==6.4.2
django-js-asset==2.0.0
filelock==3.7.1
flake8==4.0.1
identify==2.5.2
mccabe==0.6.1
nodeenv==1.7.0
Pillow==9.2.0
pkg_resources==0.0.0
platformdirs==2.5.2
pre-commit==2.20.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
PyYAML==6.0
sqlparse==0.4.2
toml==0.10.2
uuid==1.30
virtualenv==20.16.2
